Remap object IDs in TOROT texts.
The div, sentence and token IDs in TOROT texts have changed since the
texts were moved from one database to another and the IDs unfortunately
reflect the database-internal keys.

This commit remaps object IDs where this was possible to do in an
automated way, i.e. when the number of objects have remained the same
and only the IDs have changed.